Fix pr #180: Fix issue #179: Create winners view
