+++ photo/queries.py
@@ -78,44 +78,15 @@ class Query:
         self, filters: Optional[CollectionFilter] = strawberry.UNSET
     ) -> List[CollectionType]:
         queryset = Collection.objects.all()
+@strawberry.field
+    def contests_with_winners(self) -> List[ContestType]:
+        contests = Contest.objects.filter(contestsubmission__isnull=False).distinct()
+        contests = contests.order_by('-voting_draw_end')
+        return contests
 
         return strawberry_django.filters.apply(filters, queryset)
 
     @strawberry.field
-@strawberry.field
-    def contests_with_winners(self) -> List[ContestType]:
-        contests = Contest.objects.filter(winners__isnull=False).order_by('-voting_draw_end')
-        result = []
-        for contest in contests:
-            winners = []
-            for winner in contest.winners.all():
-                submission = ContestSubmission.objects.filter(contest=contest, picture__user=winner).first()
-                if submission:
-                    winners.append(
-                        WinnerType(
-                            name_first=winner.name_first,
-                            name_last=winner.name_last,
-                            submission=WinnerSubmissionType(
-                                picture=WinnerPictureType(
-                                    name=submission.picture.name,
-                                    file=submission.picture.file.url
-                                ),
-                                number_votes=submission.votes.count()
-                            )
-                        )
-                    )
-            if winners:
-                result.append(
-                    ContestType(
-                        title=contest.title,
-                        description=contest.description,
-                        prize=contest.prize,
-                        voting_draw_end=contest.voting_draw_end,
-                        winners=winners
-                    )
-                )
-        return result
-
     def contests(
         self, filters: Optional[ContestFilter] = strawberry.UNSET
     ) -> List[ContestType]:
